Move local classes to their rightful location
